requirements.txt
# arrow==1.3.0
# asgiref==3.8.1
# asttokens==3.0.0
# attrs==25.3.0
# autocommand==2.2.2
# backcall==0.2.0
# backports.tarfile==1.2.0
# beautifulsoup4==4.13.3
# bleach==6.2.0
# certifi==2025.1.31
# charset-normalizer==3.4.1
# colorama==0.4.6
# decorator==5.2.1
# defusedxml==0.7.1
# distlib==0.3.9
# Django==5.1.7
# django-cors-headers==4.7.0
# django-environ==0.12.0
# djangorestframework==3.15.2
# docopt==0.6.2
# executing==2.2.0
# fastjsonschema==2.21.1
# filelock==3.18.0
# fqdn==1.5.1
# idna==3.10
# importlib_metadata==8.0.0
# ipython==8.12.3
# ipython_pygments_lexers==1.1.1
# isoduration==20.11.0
# jaraco.collections==5.1.0
# jaraco.context==6.0.1
# jaraco.functools==4.1.0
# jaraco.text==4.0.0
# jedi==0.19.2
# Jinja2==3.1.6
# jsonpointer==3.0.0
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# jupyterlab_pygments==0.3.0
# MarkupSafe==3.0.2
# matplotlib-inline==0.1.7
# mistune==3.1.3
# more-itertools==10.6.0
# nbclient==0.10.2
# nbconvert==7.16.6
# nbformat==5.10.4
# packaging==24.2
# pandocfilters==1.5.1
# parso==0.8.4
# pickleshare==0.7.5
# pip-chill==1.0.3
# pipenv==2024.4.1
# pipreqs==0.5.0
# platformdirs==4.3.7
# prompt_toolkit==3.0.50
# psycopg-binary==3.2.6
# psycopg2==2.9.10
# pure_eval==0.2.3
# Pygments==2.19.1
# python-dateutil==2.9.0.post0
# python-decouple==3.8
# pywin32==310
# pyzmq==26.3.0
# referencing==0.36.2
# requests==2.32.3
# rpds-py==0.24.0
# setuptools==78.1.0
# six==1.17.0
# soupsieve==2.6
# sqlparse==0.5.3
# stack-data==0.6.3
# tinycss2==1.4.0
# tomli==2.0.1
# tornado==6.4.2
# traitlets==5.14.3
# types-python-dateutil==2.9.0.20241206
# typing_extensions==4.13.0
# tzdata==2025.2
# uri-template==1.3.0
# urllib3==2.3.0
# virtualenv==20.29.3
# wcwidth==0.2.13
# webcolors==24.11.1
# webencodings==0.5.1
# yarg==0.1.9
# zipp==3.21.0

#? ??? CLEAN ???
backports.tarfile==1.2.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
fqdn==1.5.1
importlib-metadata==8.0.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jaraco.collections==5.1.0
jsonpointer==3.0.0
pip-chill==1.0.3
pipenv==2024.4.1
psycopg-binary==3.2.6
psycopg2==2.9.10
python-decouple==3.8
tinycss2==1.4.0
tomli==2.0.1
uri-template==1.3.0
webcolors==24.11.1